--- Proveedores.csv
@@ -1,3 +1,0 @@
-1,000000000001234,Nicky Jam                ,55555555                                                                                                                                                   
-0,000000000006543,marta                    ,                                                                                                                                                           
-1,000000000004321,JUANCITO                 ,
