+++ files/Proveedores.csv
@@ -1,0 +1,6 @@
+1,000000000001234,Nicolas Buiatti          ,55555555                                                                                                                                                   
+0,000000000006543,marta                    ,                                                                                                                                                           
+0,000000000004321,JUANCITO                 ,                                                                                                                                                           
+1,000000000123321,prueba                   ,                                                                                                                                                            
+1,000000000011111,asdf                     ,                                                                                                                                                            
+1,000000000567654,a verte                  ,
